+++ lib/yard-sorbet/sig_to_yard.rb
@@ -75,19 +75,15 @@ module YARDSorbet
         end
       end
 
+      # @see https://www.rubydoc.info/gems/yard/file/docs/Tags.md#order-dependent-lists Order-Dependent Lists
       sig { params(node: YARD::Parser::Ruby::AstNode).returns([String]) }
-      def convert_array(node)
-        # https://www.rubydoc.info/gems/yard/file/docs/Tags.md#order-dependent-lists
+      def convert_array(node) = ["Array<(#{sequence(node.first)})>"]
 
-        if node.first.nil?
-          # Empty tuples
-          ['Array<()>']
-        else
-          # Non-empty tuples
-          member_types = node.first.children.map { convert_node(_1) }
-          sequence = member_types.map { _1.size == 1 ? _1[0] : _1.to_s.tr('"', '') }.join(', ')
-          ["Array<(#{sequence})>"]
-        end
+      sig { params(node: T.nilable(YARD::Parser::Ruby::AstNode)).returns(String) }
+      def sequence(node)
+        return '' unless node
+
+        node.children.map { convert_node(_1) }.map { _1.size == 1 ? _1[0] : _1.to_s.tr('"', '') }.join(', ')
       end
 
       sig { params(node: YARD::Parser::Ruby::AstNode).returns([String]) }