+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## main
 
+### Bug Fixes
+* [#351](https://github.com/dduugg/yard-sorbet/pull/351) Fix handling of empty tuples (fixed arrays)
+
 ## 0.9.0 (2024-06-30)
 
 ### Bug Fixes